// apps/portal-api/src/model/kanel/public/ServiceGroupUser.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ServiceGroupId } from './ServiceGroup';
import type { UserId } from './User';

/** Represents the table public.ServiceGroup_User */
export default interface ServiceGroupUser {
  group_id: ServiceGroupId;

  user_id: UserId;
}

/** Represents the initializer for the table public.ServiceGroup_User */
export interface ServiceGroupUserInitializer {
  group_id: ServiceGroupId;

  user_id: UserId;
}

/** Represents the mutator for the table public.ServiceGroup_User */
export interface ServiceGroupUserMutator {
  group_id?: ServiceGroupId;

  user_id?: UserId;
}